Extracted XmlParser
