+++ src/CodeGenerator/Generator/GeneratorHelper.php
@@ -12,6 +12,7 @@ use Nette\PhpGenerator\Method;
  * Small methods that might be useful when generating code.
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ * @author Jérémy Derussé <jeremy@derusse.com>
  *
  * @internal
  */
